Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
google-genai==0.2.2
trafilatura==1.6.3
duckduckgo-search==6.1.7
requests==2.32.3
